Feat: Add file to store helper functions used in index.js
