+++ server/index.js
@@ -4,122 +4,15 @@ const cors = require("cors");
 const wanakana = require("wanakana");
 const knex = require("./db/knex");
 const path = require("path");
+const {
+  katakanaToKanji,
+  getKanjiDefinitions,
+  combiner,
+  getKanjiData,
+} = require("./helper.js");
 
 app.use(express.json()); //req.body
 app.use(cors());
-
-function shuffle(arr) {
-  const cloneArray = [...arr];
-
-  for (let i = cloneArray.length - 1; i >= 0; i--) {
-    let rand = Math.floor(Math.random() * (i + 1));
-    // change the element of order in array
-    let tmpStorage = cloneArray[i];
-    cloneArray[i] = cloneArray[rand];
-    cloneArray[rand] = tmpStorage;
-  }
-
-  return cloneArray;
-}
-
-async function katakanaToKanji(katakanaArr) {
-  const result = {};
-  let count = 1;
-
-  for (const katakana of katakanaArr) {
-    // Get kanji character for each katakana character
-    const kanjiResponseObj = await fetch(
-      `https://kanjialive-api.p.rapidapi.com/api/public/search/advanced/?on=${katakana}`,
-      {
-        method: "GET",
-        headers: {
-          "X-RapidAPI-Key":
-            "fb204f9fc6msh5fe937040b3b47fp1ec3cfjsn0a6c2ab1a027",
-          "X-RapidAPI-Host": "kanjialive-api.p.rapidapi.com",
-        },
-      }
-    );
-
-    // Get body of response from API which will be array of kanji objects
-    const kanjiArr = await kanjiResponseObj.json();
-    // Shuffle order of kanji objects in kanji array
-    const shuffledKanjiArr = shuffle(kanjiArr);
-    // Get first 3 kanji objects of shuffled array
-    const kanjiTriplet = shuffledKanjiArr.slice(0, 3);
-
-    const kanjiCharactersArr = kanjiTriplet.map((singleKanjiObj) => {
-      return singleKanjiObj.kanji.character;
-    });
-
-    result[count++] = kanjiCharactersArr;
-  }
-
-  return result;
-}
-
-async function getKanjiDefinitions(generatedKanjiObj) {
-  const result = {};
-  let count = 1;
-
-  for (const key in generatedKanjiObj) {
-    const definitionArr = [];
-    for (const kanji of generatedKanjiObj[key]) {
-      const kanjiResponseObj = await fetch(
-        `https://kanjialive-api.p.rapidapi.com/api/public/kanji/${kanji}`,
-        {
-          method: "GET",
-          headers: {
-            "X-RapidAPI-Key":
-              "fb204f9fc6msh5fe937040b3b47fp1ec3cfjsn0a6c2ab1a027",
-            "X-RapidAPI-Host": "kanjialive-api.p.rapidapi.com",
-          },
-        }
-      );
-
-      // Get body of response from API which will be array of definition objects
-      const kanjiObj = await kanjiResponseObj.json();
-      // Get and store definition of each kanji object
-      const definition = kanjiObj.kanji.meaning.english;
-      definitionArr.push(definition);
-    }
-
-    result[count++] = definitionArr;
-  }
-  
-  return result;
-}
-
-// Function to combine kanji/definition pairs/triplets/quadruplets/etc.
-const combiner = (obj) => {
-  const result = {};
-
-  for (const key in obj) {
-    for (let i = 0; i < obj[key].length; i++) {
-      if (!result[i + 1]) {
-        result[i + 1] = [];
-      }
-      result[i + 1].push(obj[key][i]);
-    }
-  }
-
-  return result;
-};
-
-// Function to organize kanji character and definition values
-const getKanjiData = (hiraganaArr, kanjiObj, definitionObj) => {
-  return Object.keys(kanjiObj).map((key) => {
-    return {
-      kanjiName: kanjiObj[key].join(""),
-      eachKanji: kanjiObj[key].map((kanji, index) => {
-        return {
-          character: kanji,
-          hiragana: hiraganaArr[index],
-          definition: definitionObj[key][index],
-        };
-      }),
-    };
-  });
-};
 
 app.get("/api/kanji/", async (req, res) => {
   const { input: romajiName } = req.query;
@@ -137,29 +30,6 @@ app.get("/api/kanji/", async (req, res) => {
   const kanjiNames = combiner(generatedKanjiObj);
   const kanjiDefinitions = combiner(generatedDefinitionObj);
   const kanjiData = getKanjiData(hiraganaArr, kanjiNames, kanjiDefinitions);
-
-  // let resultObj = [
-  //   {
-  //     kanjiName: "区理酢",
-  //     eachKanji: [
-  //       {
-  //         character: "区",
-  //         hiragana: "く",
-  //         meanings: "district",
-  //       },
-  //       {
-  //         character: "理",
-  //         hiragana: "り",
-  //         meanings: "science",
-  //       },
-  //       {
-  //         character: "酢",
-  //         hiragana: "す",
-  //         meanings: "vinegar",
-  //       },
-  //     ],
-  //   },
-  // ];
 
   res.status(200).send(JSON.stringify({ kanjiData }));
 });
